+++ resources/displaySampleWindow.py
@@ -122,7 +122,12 @@ class displaySampleWindow(QWidget):
     #---------------------------------------------------------------------------------------------
     def closeEvent(self, event):
         self.sampleDict['Comment'] = self.textComment.toPlainText()
-        self.item.setData(0, Qt.UserRole, self.sampleDict)
+        # if WS has been removed while a Display is active 
+        try:
+            self.item.setData(0, Qt.UserRole, self.sampleDict)
+        except:
+            #print("item not available to be updated")
+            pass 
         self.open_displayWindows.pop(self.Id, None)
         event.accept()
 
@@ -130,9 +135,6 @@ class displaySampleWindow(QWidget):
 # Example usage
 if __name__ == "__main__":
 
-    def handle_item(item):
-        print('handle', item)
- 
     app = QApplication([])
 
     itemDict = {